Move resultReader to a separate file
