+++ test/converters/junitxml/junitxml.go
@@ -2,37 +2,11 @@ package junitxml
 
 import (
 	"encoding/xml"
-	"os"
 	"strings"
 
 	"github.com/bitrise-steplib/steps-deploy-to-bitrise-io/test/testreport"
 	"github.com/pkg/errors"
 )
-
-type resultReader interface {
-	ReadAll() ([]byte, error)
-}
-
-type fileReader struct {
-	Filename string
-}
-
-func (r *fileReader) ReadAll() ([]byte, error) {
-	return os.ReadFile(r.Filename)
-}
-
-type stringReader struct {
-	Contents string
-}
-
-func (r *stringReader) ReadAll() ([]byte, error) {
-	return []byte(r.Contents), nil
-}
-
-// Converter holds data of the converter
-type Converter struct {
-	results []resultReader
-}
 
 func (c *Converter) Setup(_ bool) {}
